requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.5.0
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
